integration: added a helper to run tests, upload to GCS and email
Updates #10

A helper to get us closer to integration tests that are triggered
say on every push to a repository.

The helper when run will run the tests by `make test` and if
there is any failure, it'll upload all the output to GCS, grab
a link and then compose an email that'll be sent to everyone
that needs to be notified.

It uses postmark for emailing

```shell
$ go run run+notify.go -gcp-project census-demos \
-postmark-account-token "dbfb2fd0-9c33-42cf-85ee-6d234ebca614" \
-postmark-server-token "85d8e1b7-52a1-41fb-8ac9-4eb5d450996e"
```

```shell
2018/03/28 00:25:13 Failed to run `make test` error: exit status 2
output: protoc -I src/main/proto/ src/main/proto/defs.proto
--go_out=plugins=grpc:src/main/proto
 # To use a different OpenCensus and gRPC version, use
 # -PopencensusVersion= and -PgrpcVersion=.
./gradlew installDist

FAILURE: Build failed with an exception.

* What went wrong:
Could not determine java version from '9.0.4'.

* Try:
Run with --stacktrace option to get the stack trace. Run with --info or
--debug option to get more log output.
make: *** [gradlew_install] Error 1
exit status 1
```