documentation for the protocol and example codes for the arduino (#46)
Design a protocol based on the serial communication between Raspberry Pi and other IoT platforms. Any other MCU could upload their data to the Pi and export the data to the back-end server if it follows the protocol.  More details about the protocol could be found in the arduino/protocol.txt.  There is also an example code file for the Arduino UNO in the /iot/arduino directory.