Start adding a java.util.logging demo for opencensus-contrib-log-correlation-stackdriver. (#52)
This commit adds a demo for log correlation using OpenCensus, java.util.logging,
and Stackdriver.  The Java application is the same as the application in #51,
except that its log statements use java.util.logging instead of SLF4J.

The demo can be run with `./gradlew run` in the
`java/log_correlation/stackdriver/java_util_logging` directory.