Add a demo for io.opencensus:opencensus-contrib-log-correlation-log4j.
This commit adds a demo for log correlation using OpenCensus and Log4j.  The
code was copied from
https://github.com/census-instrumentation/opencensus-java/pull/1382 and modified
to be built as an independent project.  It currently depends on a snapshot
version of opencensus-java, since opencensus-contrib-log-correlation-log4j
hasn't been released yet.  See the readme for instructions for running the demo.